break up heatmap index:
- the supply-defaults, calc, plot, style and hoverPoints steps
  each get a file of their own.
- helpers convertColumnXYZ, hasColumns and maxRowLength are now modules
  required into module files.
